Okay, we've done a few things.
First we fixed the scheduler, because the new EVec buffer was acting up
(flushing data before reading)

Secondly, we're looking at 2 new things we're going to add.
First we're actually going to seperate DelayQueue from Schedule
and place it outside of CronPool so that we may start chaining pools
together, this will have on pool feed its results into another. The
DelayQueue will be placed in some management struct in the future.

So DelayQueue is temporarily depreciated, but will be added again in
some kind of manager. I'll probably make a new branch for this.